+++ feature/src/main/java/team/aliens/dms/android/feature/resetpassword/ResetPasswordEnterEmailVerificationCodeScreen.kt
@@ -1,4 +1,4 @@
-package team.aliens.dms.android.feature.signup
+package team.aliens.dms.android.feature.resetpassword
 
 import android.os.CountDownTimer
 import androidx.compose.foundation.layout.Column
@@ -27,11 +27,10 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.ramcosta.composedestinations.annotation.Destination
 import team.aliens.dms.android.core.designsystem.ButtonDefaults
 import team.aliens.dms.android.core.designsystem.ContainedButton
-import team.aliens.dms.android.core.designsystem.DmsIcon
-import team.aliens.dms.android.core.designsystem.Scaffold
 import team.aliens.dms.android.core.designsystem.DmsTheme
 import team.aliens.dms.android.core.designsystem.DmsTopAppBar
 import team.aliens.dms.android.core.designsystem.LocalToast
+import team.aliens.dms.android.core.designsystem.Scaffold
 import team.aliens.dms.android.core.designsystem.TextButton
 import team.aliens.dms.android.core.designsystem.VerificationCodeInput
 import team.aliens.dms.android.core.designsystem.VerificationCodeInputDefaults
@@ -44,7 +43,7 @@ import team.aliens.dms.android.core.ui.horizontalPadding
 import team.aliens.dms.android.core.ui.startPadding
 import team.aliens.dms.android.core.ui.topPadding
 import team.aliens.dms.android.feature.R
-import team.aliens.dms.android.feature.signup.navigation.SignUpNavigator
+import team.aliens.dms.android.feature.resetpassword.navigation.ResetPasswordNavigator
 
 // Millisecond * Second * Minutes
 private const val TIMER_TOTAL_SECONDS: Long = 1000 * 60 * 3
@@ -53,17 +52,19 @@ private const val TIMER_INTERVAL: Long = 1000
 @OptIn(ExperimentalMaterial3Api::class)
 @Destination
 @Composable
-internal fun SignUpEnterEmailVerificationCodeScreen(
+internal fun ResetPasswordEnterEmailVerificationCodeScreen(
     modifier: Modifier = Modifier,
-    navigator: SignUpNavigator,
-    viewModel: SignUpViewModel,
+    navigator: ResetPasswordNavigator,
+    viewModel: ResetPasswordViewModel,
 ) {
     val uiState by viewModel.stateFlow.collectAsStateWithLifecycle()
-    var timerText: String by remember { mutableStateOf("") }
     val toast = LocalToast.current
     val context = LocalContext.current
+    var timerText: String by remember { mutableStateOf("") }
+
     var isVerificationInputAvailable by remember { mutableStateOf(true) }
     var isSessionExpired by remember { mutableStateOf(false) }
+
     val timer: CountDownTimer = remember {
         object : CountDownTimer(TIMER_TOTAL_SECONDS, TIMER_INTERVAL) {
             override fun onTick(millisUntilFinished: Long) {
@@ -75,14 +76,8 @@ internal fun SignUpEnterEmailVerificationCodeScreen(
             override fun onFinish() {
                 isVerificationInputAvailable = false
                 isSessionExpired = true
-                viewModel.postIntent(SignUpIntent.UpdateEmailVerificationCode(value = ""))
+                viewModel.postIntent(ResetPasswordIntent.UpdateEmailVerificationCode(value = ""))
             }
-        }
-    }
-
-    LaunchedEffect(uiState.emailVerificationCode) {
-        if (uiState.emailVerificationCode.length == SignUpViewModel.EMAIL_VERIFICATION_CODE_LENGTH) {
-            viewModel.postIntent(SignUpIntent.CheckEmailVerificationCode)
         }
     }
 
@@ -92,16 +87,16 @@ internal fun SignUpEnterEmailVerificationCodeScreen(
 
     viewModel.sideEffectFlow.collectInLaunchedEffectWithLifecycle { sideEffect ->
         when (sideEffect) {
-            SignUpSideEffect.EmailVerificationCodeChecked -> {
+            ResetPasswordSideEffect.EmailVerificationCodeChecked -> {
                 timer.cancel()
-                navigator.openSetId()
+                navigator.openResetPasswordSetPassword()
             }
 
-            SignUpSideEffect.EmailVerificationCodeIncorrect -> toast.showErrorToast(
+            ResetPasswordSideEffect.EmailVerificationCodeIncorrect -> toast.showErrorToast(
                 message = context.getString(R.string.sign_up_enter_email_verification_code_error_verification_code_incorrect),
             )
 
-            SignUpSideEffect.EmailVerificationSessionReset -> {
+            ResetPasswordSideEffect.EmailVerificationSessionReset -> {
                 isVerificationInputAvailable = true
                 with(timer) {
                     cancel()
@@ -112,11 +107,11 @@ internal fun SignUpEnterEmailVerificationCodeScreen(
                 )
             }
 
-            SignUpSideEffect.EmailVerificationSessionResetFailed -> toast.showErrorToast(
+            ResetPasswordSideEffect.EmailVerificationSessionResetFailed -> toast.showErrorToast(
                 message = context.getString(R.string.sign_up_enter_email_verification_code_error_cannot_resend_verification_code),
             )
 
-            else -> {/* explicit blank */
+            else -> { // * explicit blank *//*
             }
         }
     }
@@ -127,9 +122,9 @@ internal fun SignUpEnterEmailVerificationCodeScreen(
             DmsTopAppBar(
                 title = {},
                 navigationIcon = {
-                    IconButton(onClick = navigator::popUpToEnterEmail) {
+                    IconButton(onClick = navigator::openSignIn) {
                         Icon(
-                            painter = painterResource(id = DmsIcon.Back),
+                            painter = painterResource(id = R.drawable.ic_baseline_arrow_back_24),
                             contentDescription = stringResource(id = R.string.top_bar_back_button),
                         )
                     }
@@ -157,10 +152,10 @@ internal fun SignUpEnterEmailVerificationCodeScreen(
                 modifier = Modifier
                     .fillMaxWidth()
                     .horizontalPadding(),
-                totalLength = SignUpViewModel.EMAIL_VERIFICATION_CODE_LENGTH,
+                totalLength = ResetPasswordViewModel.EMAIL_VERIFICATION_CODE_LENGTH,
                 text = uiState.emailVerificationCode,
                 onValueChange = { verificationCode ->
-                    viewModel.postIntent(SignUpIntent.UpdateEmailVerificationCode(value = verificationCode))
+                    viewModel.postIntent(ResetPasswordIntent.UpdateEmailVerificationCode(value = verificationCode))
                 },
                 supportingText = {
                     VerificationCodeInputDefaults.SupportingText(
@@ -196,7 +191,7 @@ internal fun SignUpEnterEmailVerificationCodeScreen(
                 modifier = Modifier
                     .fillMaxWidth()
                     .horizontalPadding(),
-                onClick = { viewModel.postIntent(SignUpIntent.ResetEmailVerificationSession) },
+                onClick = { viewModel.postIntent(ResetPasswordIntent.ResetEmailVerificationSession) },
                 colors = ButtonDefaults.textGrayButtonColors(),
             ) {
                 Text(text = stringResource(id = R.string.sign_up_enter_email_verification_code_resend_verification_code))
@@ -207,8 +202,8 @@ internal fun SignUpEnterEmailVerificationCodeScreen(
                     .fillMaxWidth()
                     .horizontalPadding()
                     .bottomPadding(),
-                onClick = { viewModel.postIntent(SignUpIntent.CheckEmailVerificationCode) },
-                enabled = uiState.emailVerificationCode.length == SignUpViewModel.EMAIL_VERIFICATION_CODE_LENGTH,
+                onClick = { viewModel.postIntent(ResetPasswordIntent.CheckEmailVerificationCode) },
+                enabled = uiState.emailVerificationCode.length == ResetPasswordViewModel.EMAIL_VERIFICATION_CODE_LENGTH,
             ) {
                 Text(text = stringResource(id = R.string.verify))
             }
